Using extract delegate refactoring
