+++ src/ifs/InsuranceCalculator.java
@@ -1,0 +1,21 @@
+package ifs;
+
+public class InsuranceCalculator {
+
+    private final ifs.InsuranceStrategyVeryHigh insuranceStrategyVeryHigh = new ifs.InsuranceStrategyVeryHigh();
+
+    public double calculateInsurance(double income) {
+
+
+        if (income <= 10000) {
+            return income * 0.365;
+        } else if (income <= 30000) {
+            return (income - 10000) * 0.2 + 35600;
+        } else if (income <= 60000) {
+            return (income - 30000) * 0.1 + 76500;
+        } else {
+            return insuranceStrategyVeryHigh.calculateInsuranceVeryHigh(income);
+        }
+
+    }
+}
