+++ test/IfElseDemoTest.java
@@ -1,3 +1,4 @@
+import ifs.InsuranceCalculator;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
